demonstration for bug #78
